Remove admin database access requirement
Hi @JamesCropcho, @smgardner59,

I needed to do a bit more to get variety working with a user that only had access to the target db:
```
mongo -u USER -p PASS TARGET_DB --eval "var collection = 'COLLECTION'" variety.js
```

This fixes #40 at the expense of error reporting when the specified database is empty or does not exist.
Lines 29-46 needed to be removed because they rely on an admin function. This also deprecates lines  21 and 22. Code review requested, because I just did this as a quick fix.